Make drawer work on all tabs
